moved text input to separate component
